common.utils: autmatic refactor: PostgreSQL
Move `get_postgres`, `with_postgres`, `pick_random_row` and
`escape_like` to `common.postgres`.
