+++ common/utils.py
@@ -11,7 +11,6 @@ import time
 
 import flask
 import irc.client
-import psycopg2
 import werkzeug.datastructures
 
 from common import config
@@ -274,18 +273,6 @@ def immutable(obj):
 		return obj
 
 
-def get_postgres():
-	return psycopg2.connect(config.config["postgres"])
-
-def with_postgres(func):
-	"""Decorator to pass a PostgreSQL connection and cursor to a function"""
-	@functools.wraps(func)
-	def wrapper(*args, **kwargs):
-		with get_postgres() as conn, conn.cursor() as cur:
-			return func(conn, cur, *args, **kwargs)
-	return wrapper
-
-
 def sse_send_event(endpoint, event=None, data=None, event_id=None):
 	if not os.path.exists(config.config['eventsocket']):
 		return
@@ -307,17 +294,6 @@ def sse_send_event(endpoint, event=None, data=None, event_id=None):
 def ucfirst(s):
 	return s[0].upper() + s[1:]
 
-
-def pick_random_row(cur, query, params = ()):
-	" Return a random row of a SELECT query. "
-	# CSE - common subexpression elimination, an optimisation Postgres doesn't do
-	cur.execute("CREATE TEMP TABLE cse AS " + query, params)
-	if cur.rowcount <= 0:
-		return None
-	cur.execute("SELECT * FROM cse OFFSET %s LIMIT 1", (random.randrange(cur.rowcount), ))
-	row = cur.fetchone()
-	cur.execute("DROP TABLE cse")
-	return row
 
 def weighted_choice(options):
 	"""
@@ -352,5 +328,3 @@ def weighted_choice(options):
 	return values[left]
 
 
-def escape_like(s):
-	return s.replace('\\', '\\\\').replace('%', '\\%').replace('_', '\\_')
